be: stock: models: change volume column type to BigIntegerField

+++ backend/stock/models.py
@@ -20,7 +20,7 @@ class Candle(models.Model):
     high = models.FloatField()
     low = models.FloatField()
     close = models.FloatField()
-    volume = models.IntegerField(default=0)
+    volume = models.PositiveBigIntegerField(default=0)
 
     date = models.DateField()
